fix: 修复多项UB/数据竞争/跨平台一致性问题
- map.hpp: value_type改为pair<Key,T>，消除reinterpret_cast和const_cast UB
- zip_extract.hpp: uint8_t→uint32_t显式转换，避免有符号左移溢出
- human_approval.hpp: status_改为std::atomic，修复多线程数据竞争
- chat_repl.cpp: 信号处理器使用std::atomic指针，修复shared_ptr数据竞争
- spinner.hpp: label_加mutex保护，修复多线程数据竞争
- string.hpp/cpp: 添加非const data()重载，消除const_cast
- platform.cpp: sysctl后检查size，避免memcpy读未初始化数据
- lock_free.hpp: spinlock_使用mutable替代const_cast
- crypto.cpp: NULL→nullptr
- 标准化平台宏: _WIN32→BEN_GEAR_PLATFORM_*

+++ src/capabilities/skill/zip_extract.hpp
@@ -26,7 +26,7 @@ constexpr uint16_t kMethodDeflated = 8;
 inline uint16_t read_u16_le(const uint8_t* p) { return p[0] | (p[1] << 8); }
 /// 读取小端 32-bit（内联，热点路径）
 inline uint32_t read_u32_le(const uint8_t* p) {
-    return p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24);
+    return p[0] | (static_cast<uint32_t>(p[1]) << 8) | (static_cast<uint32_t>(p[2]) << 16) | (static_cast<uint32_t>(p[3]) << 24);
 }
 
 /// 使用 zlib 解压 deflate 数据
